pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-cartoon-cam"
version = "0.1.0"
description = "Real-time local camera stylization for cartoon/sketch effects on macOS."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "ai-cartoon-cam contributors" }
]
keywords = ["opencv", "virtual-camera", "cartoon", "obs", "macos"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "License :: OSI Approved :: MIT License",
  "Operating System :: MacOS",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Topic :: Multimedia :: Video",
  "Topic :: Scientific/Engineering :: Image Processing"
]
dependencies = [
  "opencv-python>=4.10.0",
  "numpy>=1.26.0",
  "PyYAML>=6.0.1"
]

[project.optional-dependencies]
virtualcam = [
  "pyvirtualcam>=0.12.0"
]
dev = [
  "pytest>=8.0.0",
  "ruff>=0.5.0",
  "mypy>=1.10.0"
]

[project.scripts]
ai-cartoon-cam = "aicartooncam.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
